+++ tests/phpstan/subscription-product-type-stubs.php
@@ -1,0 +1,4 @@
+<?php
+
+class WC_Product_Simple {
+}